remove mpi header from context for compilation with Cray mpi
